requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.1.0
plotly>=5.17.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
